Refactor customizable input variable handling
Change code to always correctly pass through customizable input
variables to the cluster module. We also move the definition of the
provider-specific input variables for the whole Terraform environment
into `component/main.jsonnet` since there's no reason for them to be
user-configurable.

With some use of object-comprehension we ensure that any input variables
which are defined in `input_vars.<provider>` are passed through to
identically named input variables of the cluster module.

This should address the failing test for the Exoscale provider.