Provide wrapper script to configure Terraform execution environment
The wrapper script needs two stages, since we currently want to execute
some preparatory commands as root (e.g. install `curl`), before
switching to an unprivileged user for actually running Terraform.

Additionally, we set `GIT_ASKPASS` to the provided `git-askpass.sh`
script, so we can support Git-over-HTTPS for private repositories.

Using Git-over-HTTPS removes the need for providing an execution
environment for SSH and means we only need to provide a token/password
instead of a SSH private key in the Terraform CI environment.